Max of min for every window size
Maximum of minimums of every window size in a given array
Given an integer array arr[] of size n, the task is to find the maximum of the minimums for every window size in the given array, where the window size ranges from 1 to n.

Example:

Input: arr[] = [10, 20, 30]
Output: [30, 20, 10]
Explanation:
First element in output indicates maximum of minimums of all windows of size 1. Minimums of windows of size 1 are [10], [20], [30]. Maximum of these minimums are 30 and similarly other outputs can be computed

Input: arr[] = [10, 20, 30, 50, 10, 70, 30]
Output: [70, 30, 20, 10, 10, 10, 10]
Explanation: The first element in the output indicates the maximum of minimums of all windows of size 1. 
Minimums of windows of size 1 are [10], [20], [30], [50], [10], [70] and [30]. 
Maximum of these minimums is 70
The second element in the output indicates the maximum of minimums of all windows of size 2. 
Minimums of windows of size 2 are [10], [20], [30], [10], [10], and [30]. 
Maximum of these minimums is 30
The third element in the output indicates the maximum of minimums of all windows of size 3. 
Minimums of windows of size 3 are [10], [20], [10], [10] and [10]. 
Maximum of these minimums is 20
Similarly, other elements of output are computed. 


Using Stack - O(n) Time and O(n) Space
The idea is to find the next smaller and previous smaller of each element and update the maximum of window with size as the difference in their indices. This problem is mainly a variation of Largest Area in a Histogram.

Below is the implementation of the above approach:

Initialize res[] to store maximum minimums and len[] to store window sizes. Use a stack s for efficient processing.
Traverse the array to determine window sizes. For each element, pop larger or equal elements from the stack, calculate their window sizes, and update len[]. Push the current index onto the stack.
Process remaining stack elements to determine their window sizes using the right boundary.
Populate res[] using len[] and arr[], storing the maximum minimum for each window size.
Update res[] to ensure values correctly reflect the maximum of the minimums.
Return res[], containing the maximum of minimums for every window size.